fix: user with no verification code deletion fix
1. added code to catch JwtAuthFilter exceptions, and build a dedicated
   exception handler class for security exceptions

2. altered DB contraint that forced the naming of is_active on the
   isDeleted field

3. changed the timestamp appended to the email of the delted user to be
   the date time of deletion

4. handled the case of a user with no verification code in delete
   verificaiton code method, to allow it be usable inside softDeleteUser
   method, because in the case of deleting a user with no verification
   code the said method, it throwed a null pointer exception

+++ src/main/java/com/na7ki/backend/domain/user/verification_code/VerificationCodeService.java
@@ -40,6 +40,10 @@ public class VerificationCodeService {
 
     public void deleteCode (User associatedUser) {
         VerificationCode associatedCode = associatedUser.getVerificationCode();
+        if (associatedCode == null)
+        {
+            return;
+        }
         associatedUser.setVerificationCode(null);
         associatedCode.setUser(null);
         verificationCodeRepository.delete(associatedCode);
